fix(hooks): correct jq-null false block + base-tracking gh order; document guard
Two review catches on the merged-branch guard (slop-mop#287):

- jq null = false block (high): `--jq '.[0].number'` prints the literal
  "null" for an empty merged-PR list, which is non-empty and blocked every
  open branch as "PR #null" once gh is installed/authed. Use
  `.[0].number // empty` so "no merged PR" yields an empty string.
  (The welcome-to-willville original has this same bug.)
- gh ran before the base-tracking skip (medium): a fresh base-tracking
  branch reusing an old merged branch's name got false-blocked by the
  stale PR. Move the base-tracking early-return ahead of the gh check.

Tests: a fake `gh` that applies the --jq filter exercises the merged-PR
path faithfully — empty result allows (jq fix), real merged PR blocks,
base-tracking branch with a stale PR allows (reorder fix). Hook tests
split into tests/unit/test_git_hooks.py to keep test_sm_cli.py under the
sprawl limit.

Docs: README "Use with pre-commit" now explains the commit-time guard.
