+++ tests/unit/test_sm_cli.py
@@ -14,11 +14,6 @@ from slopmop.cli.buff import cmd_buff
 from slopmop.cli.help import cmd_help
 from slopmop.cli.hooks import (
     SB_HOOK_MARKER,
-    _generate_hook_script,
-    _generate_merged_branch_guard,
-    _generate_pre_push_hook_script,
-    _get_git_hooks_dir,
-    _parse_hook_info,
     cmd_commit_hooks,
 )
 from slopmop.cli.init import prompt_user, prompt_yes_no
@@ -244,179 +239,6 @@ class TestCmdHelp:
         assert result == 0
         captured = capsys.readouterr()
         assert "Python Tests" in captured.out
-
-
-class TestGitHooksFunctions:
-    """Tests for git hooks helper functions."""
-
-    def test_get_git_hooks_dir(self, tmp_path):
-        """Returns hooks dir for git repo."""
-        (tmp_path / ".git").mkdir()
-        result = _get_git_hooks_dir(tmp_path)
-        assert result == tmp_path / ".git" / "hooks"
-
-    def test_get_git_hooks_dir_not_git(self, tmp_path):
-        """Returns None for non-git directory."""
-        result = _get_git_hooks_dir(tmp_path)
-        assert result is None
-
-    def test_generate_hook_script(self):
-        """Generates valid hook script with swab verb."""
-        script = _generate_hook_script("swab")
-        assert "sm swab" in script
-        assert "MANAGED BY SLOP-MOP" in script
-        # Should use PATH-based sm lookup
-        assert "command -v sm" in script
-        # Should write structured output for LLM consumption
-        assert "--swabbing-timeout 0" in script
-        assert "--json-file .slopmop/last_swab.json" in script
-        assert "--json --output-file" not in script
-        assert "Structured results:" in script
-        assert "mkdir -p .slopmop" in script
-
-    def test_generate_hook_script_direct_verb(self):
-        """Generates hook script when given a verb directly."""
-        script = _generate_hook_script("scour")
-        assert "sm scour" in script
-        assert "# Command: sm scour" in script
-        assert "--swabbing-timeout 0" in script
-        assert "--json-file .slopmop/last_scour.json" in script
-
-    def test_generate_pre_push_hook_script(self):
-        """Generates pre-push merged-branch guard hook script."""
-        script = _generate_pre_push_hook_script()
-        assert "# Command: merged-branch-guard" in script
-        assert "gh pr list" in script
-        assert "--state merged" in script
-        assert "You're missing some context." in script
-        assert "sync against main, checkout a new branch, and open a new PR." in script
-        # Guard must inspect the refs Git passes on stdin, not just HEAD, so a
-        # push that names a branch other than the current checkout is covered.
-        assert "while read -r local_ref local_sha remote_ref remote_sha" in script
-        assert "refs/heads/*) branch=${local_ref#refs/heads/}" in script
-        # Deletions (all-zero local sha) write nothing and must be skipped.
-        assert "0000000000000000000000000000000000000000" in script
-        assert "git symbolic-ref" not in script
-
-    def test_precommit_hook_embeds_merged_branch_guard(self):
-        """The pre-commit hook guards a merged/deleted branch before swab runs."""
-        script = _generate_hook_script("swab")
-        assert "merged/deleted-branch guard" in script
-        # Guard must run BEFORE the swab invocation, or work piles onto a dead
-        # branch before the gate even checks. Anchor on the real invocation
-        # (with --swabbing-timeout), not the "# Command:" comment.
-        assert script.index("\n_sm_merged_branch_guard\n") < script.index(
-            "sm swab --porcelain --swabbing-timeout"
-        )
-        # The three detection strategies, strongest first.
-        assert "gh pr list --head" in script and "--state merged" in script
-        assert "git ls-remote --heads" in script
-        assert "git merge-base --is-ancestor HEAD" in script
-        # Allow-lists: integration branches and never-pushed (no upstream).
-        assert "main|master|develop" in script
-        assert "@{upstream}" in script
-        # Escape hatch surfaced to the user.
-        assert "git commit --no-verify" in script
-
-    def test_precommit_hook_is_valid_posix_sh(self):
-        """The generated hook (guard + swab) must parse as POSIX sh."""
-        import subprocess
-
-        for verb in ("swab", "scour"):
-            script = _generate_hook_script(verb)
-            result = subprocess.run(
-                ["sh", "-n", "/dev/stdin"],
-                input=script,
-                text=True,
-                capture_output=True,
-            )
-            assert result.returncode == 0, f"{verb}: {result.stderr}"
-
-    def _run_guard(self, cwd) -> int:
-        """Write the guard to a temp script and run it in cwd; return exit code."""
-        import subprocess
-
-        guard = "#!/bin/sh\n" + _generate_merged_branch_guard() + "\nexit 0\n"
-        script = cwd / ".guard.sh"
-        script.write_text(guard)
-        return subprocess.run(["sh", str(script)], cwd=str(cwd)).returncode
-
-    def _git(self, cwd, *args) -> None:
-        import subprocess
-
-        subprocess.run(["git", *args], cwd=str(cwd), check=True, capture_output=True)
-
-    def test_guard_blocks_branch_merged_into_default(self, tmp_path):
-        """A branch fully contained in a moved-on origin/main is blocked."""
-        import subprocess
-
-        remote = tmp_path / "remote.git"
-        subprocess.run(["git", "init", "-q", "--bare", str(remote)], check=True)
-        work = tmp_path / "work"
-        subprocess.run(["git", "clone", "-q", str(remote), str(work)], check=True)
-        self._git(work, "config", "user.email", "t@t.t")
-        self._git(work, "config", "user.name", "t")
-        self._git(work, "commit", "-q", "--allow-empty", "-m", "init")
-        self._git(work, "branch", "-M", "main")
-        self._git(work, "push", "-q", "-u", "origin", "main")
-        self._git(work, "remote", "set-head", "origin", "main")
-        # feature branch, pushed (own tracking ref), then merged into main
-        self._git(work, "checkout", "-q", "-b", "feat/x")
-        self._git(work, "commit", "-q", "--allow-empty", "-m", "w")
-        self._git(work, "push", "-q", "-u", "origin", "feat/x")
-        self._git(work, "checkout", "-q", "main")
-        self._git(work, "merge", "-q", "--no-ff", "feat/x", "-m", "merge")
-        self._git(work, "push", "-q", "origin", "main")
-        self._git(work, "checkout", "-q", "feat/x")
-        assert self._run_guard(work) == 1  # blocked
-
-    def test_guard_allows_fresh_open_branch(self, tmp_path):
-        """An open branch not yet merged is allowed through."""
-        import subprocess
-
-        remote = tmp_path / "remote.git"
-        subprocess.run(["git", "init", "-q", "--bare", str(remote)], check=True)
-        work = tmp_path / "work"
-        subprocess.run(["git", "clone", "-q", str(remote), str(work)], check=True)
-        self._git(work, "config", "user.email", "t@t.t")
-        self._git(work, "config", "user.name", "t")
-        self._git(work, "commit", "-q", "--allow-empty", "-m", "init")
-        self._git(work, "branch", "-M", "main")
-        self._git(work, "push", "-q", "-u", "origin", "main")
-        self._git(work, "remote", "set-head", "origin", "main")
-        self._git(work, "checkout", "-q", "-b", "feat/open")
-        self._git(work, "commit", "-q", "--allow-empty", "-m", "w")
-        self._git(work, "push", "-q", "-u", "origin", "feat/open")
-        assert self._run_guard(work) == 0  # allowed
-
-    def test_guard_allows_integration_branch(self, tmp_path):
-        """Committing directly on main is never blocked by the guard."""
-        import subprocess
-
-        subprocess.run(["git", "init", "-q", str(tmp_path)], check=True)
-        self._git(tmp_path, "config", "user.email", "t@t.t")
-        self._git(tmp_path, "config", "user.name", "t")
-        self._git(tmp_path, "commit", "-q", "--allow-empty", "-m", "init")
-        self._git(tmp_path, "branch", "-M", "main")
-        assert self._run_guard(tmp_path) == 0
-
-    def test_parse_hook_info_new_format(self):
-        """Parses new-format hook info (Command: sm verb)."""
-        content = """# MANAGED BY SLOP-MOP
-#!/bin/sh
-# Command: sm swab
-sm swab
-"""
-        result = _parse_hook_info(content)
-        assert result is not None
-        assert result["verb"] == "swab"
-        assert result["managed"] is True
-
-    def test_parse_hook_info_not_managed(self):
-        """Returns None for non-managed hook."""
-        content = "#!/bin/sh\necho hello"
-        result = _parse_hook_info(content)
-        assert result is None
 
 
 class TestCmdCommitHooks:
